+++ modules/core/03-connection/types/errors.go
@@ -16,4 +16,5 @@ var (
 	ErrInvalidVersion                = errorsmod.Register(SubModuleName, 9, "invalid connection version")
 	ErrVersionNegotiationFailed      = errorsmod.Register(SubModuleName, 10, "connection version negotiation failed")
 	ErrInvalidConnectionIdentifier   = errorsmod.Register(SubModuleName, 11, "invalid connection identifier")
+	ErrRedundantHandshake            = errorsmod.Register(SubModuleName, 12, "redundant handshake attempt")
 )