+++ CHANGELOG.md
@@ -153,8 +153,10 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 * (06-solomachine) [\#2744](https://github.com/cosmos/ibc-go/pull/2744) `Misbehaviour.ValidateBasic()` now only enforces that signature data does not match when the signature paths are different.
 * (06-solomachine) [\#2748](https://github.com/cosmos/ibc-go/pull/2748) Adding sentinel value for header path in 06-solomachine.
-* (apps/29-fee) [\#2942](https://github.com/cosmos/ibc-go/pull/2942) Check `x/bank` send enabled before escrowing fees.
-* (core/04-channel) [\#3009](https://github.com/cosmos/ibc-go/pull/3009) Change check to disallow optimistic sends.
+
+* (core/03-connection) [\#1723](https://github.com/cosmos/ibc-go/pull/1723) Removed ability for multiple TRY messages to succeed for a single INIT message in connection handshake. The first successful TRY message will generate the connectionID for a given counterparty's INIT attempt and all subsequent TRY messages with the same counterparty will fail.
+* (core/04-channel) [#3009](https://github.com/cosmos/ibc-go/pull/3009) Change check to disallow optimistic sends.
+
 
 ### Improvements